move Lakeshore425.py
